nvme: convert L2P tbl to 64b addressing. (& fix range error)
Converted from 32b entries to 64b entries. Also fixed an erroneous
range check error in update_l2p_tbl.

+++ hw/block/nvme.h
@@ -827,7 +827,7 @@ typedef struct NvmeNamespace {
     uint64_t        meta_start_offset;
     uint64_t        tbl_dsk_start_offset;
     uint32_t        tbl_entries;
-    uint32_t        *tbl;
+    uint64_t        *tbl;
 } NvmeNamespace;
 
 #define TYPE_NVME "nvme"
